Fixing the test spec files to incorporate the latest changes to the services
